Pin npm dependency chalk to 1.1.3

+++ docker/setup_npm.sh
@@ -1,5 +1,5 @@
 #!/bin/bash -e
-ln -s /usr/bin/nodejs /usr/bin/node
+[ -e /usr/bin/node ] || ln -s /usr/bin/nodejs /usr/bin/node
 npm install
 npm install gulp -g
 gem install sass
